requirements-dev.txt
# Non-pinned requirements file
black
pylama
yamllint
bandit
requests
pandas
python-dotenv
mypy
nox
types-python-dateutil
types-requests
data-science-types